% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arimaf}
\alias{arimaf}
\title{Title}
\usage{
arimaf(x, h = 5, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{

after::arimaf(AirPassengers, h=7)

after::arimaf(Nile, h=6)

after::arimaf(WWWusage, h=8)

after::arimaf(USAccDeaths, h=10)

}
