% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_forecast_mts_interface.R
\name{bag_ridge2f}
\alias{bag_ridge2f}
\title{Title}
\usage{
bag_ridge2f(
  x,
  B = 10,
  agg = c("mean", "median"),
  lags = 1,
  nb_hidden = 5,
  nodes_sim = c("sobol", "halton", "unif"),
  activ = c("relu", "sigmoid", "tanh", "leakyrelu", "elu", "linear"),
  hidden_layer_bias = FALSE,
  verbose = TRUE,
  col_sample = 0.5,
  row_sample = 1,
  a = 0.01,
  lambda_1 = 0.1,
  lambda_2 = 0.1,
  h = 5
)
}
\arguments{
\item{h}{}
}
\description{
Title
}
