% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ml_dynrm.R
\name{dynrm_fit}
\alias{dynrm_fit}
\title{Fit a dynrm}
\usage{
dynrm_fit(
  y,
  p,
  P = 1,
  xreg = NULL,
  fit_func = stats::lm.fit,
  predict_func = predict.lm,
  fit_params = NULL,
  lambda = NULL,
  alpha = NULL,
  scale_inputs = TRUE,
  x = y,
  centers = NULL,
  seed = 123,
  ...
)
}
\arguments{
\item{y}{lorem ipsum}

\item{p}{lorem ipsum}

\item{P}{lorem ipsum}

\item{xreg}{lorem ipsum}

\item{fit_func}{lorem ipsum}

\item{predict_func}{lorem ipsum}

\item{fit_params}{lorem ipsum}

\item{lambda}{lorem ipsum}

\item{scale_inputs}{lorem ipsum}

\item{x}{lorem ipsum}

\item{centers}{lorem ipsum}

\item{seed}{lorem ipsum}

\item{...}{lorem ipsum}
}
\description{
Fit a dynrm
}
