% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ml_dynrm.R
\name{dynrm_predict}
\alias{dynrm_predict}
\title{Predict from dynrm}
\usage{
dynrm_predict(
  out,
  h = ifelse(out$m > 1, 2 * out$m, 10),
  level = c(80, 95),
  fan = FALSE,
  xreg = NULL,
  lambda = out$lambda,
  ci = c("gaussian", "E", "A", "T", "garch"),
  ...
)
}
\arguments{
\item{out}{lorem ipsum}

\item{h}{lorem ipsum}

\item{level}{lorem ipsum}

\item{fan}{lorem ipsum}

\item{xreg}{lorem ipsum}

\item{lambda}{lorem ipsum}

\item{ci}{lorem ipsum}

\item{...}{lorem ipsum}
}
\description{
Predict from dynrm
}
