% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_forecast_after_interface.R
\name{dynrmf}
\alias{dynrmf}
\title{Title}
\usage{
dynrmf(
  y,
  h = 5,
  level = c(80, 95),
  fit_func = fit_ridge,
  predict_func = predict_ridge,
  ...
)
}
\arguments{
\item{y}{lorem ipsum}

\item{h}{lorem ipsum}

\item{level}{lorem ipsum}

\item{fit_func}{lorem ipsum}

\item{predict_func}{lorem ipsum}

\item{...}{}
}
\description{
Title
}
\examples{

par(mfrow=c(3, 2))
plot(dynrmf(USAccDeaths, h=10, level=c(80, 90, 95, 99)))
plot(dynrmf(AirPassengers, h=10, level=c(80, 90, 95, 99)))
plot(dynrmf(lynx, h=10, level=c(80, 90, 95, 99)))
plot(dynrmf(WWWusage, h=10, level=c(80, 90, 95, 99)))
plot(dynrmf(Nile, h=10, level=c(80, 90, 95, 99)))
plot(dynrmf(fdeaths, h=10, level=c(80, 90, 95, 99)))


}
