% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{eatf}
\alias{eatf}
\title{combined ets-arima-theta forecasts}
\usage{
eatf(
  y,
  h = 5,
  level = c(80, 95),
  method = c("E", "A", "T", "EAT"),
  weights = rep(1/3, 3),
  ci = c("E", "A", "T"),
  ...
)
}
\arguments{
\item{y}{a univariate time series}

\item{h}{number of periods for forecasting}

\item{level}{confidence levels for prediction intervals}

\item{method}{forecasting method: "E": \code{forecast::ets};
"A": \code{forecast::auto.arima}; "T": \code{forecast::thetaf};
 or "EAT" for the combination of the three methods (with weights)}

\item{weights}{weights of each method, in method 'EAT'}

\item{...}{additional parameters to be passed to \code{forecast::ets},
\code{forecast::auto.arima}, \code{forecast::thetaf} and
\code{forecast::forecast}}
}
\value{
a list with point forecasts and prediction intervals; returns
from \code{forecast::ets}, \code{forecast::auto.arima},
\code{forecast::theta}, only point forecasts for eat

\code{mean} point forecasts as a time series
}
\description{
combined ets, arima, and theta (eat) forecasting
}
\details{
ensemble forecasts obtained from \code{forecast::ets},
\code{forecast::auto.arima} and \code{forecast::theta}
}
\examples{

require(forecast)

print(after::eatf(WWWusage, method = "EAT",
weights = c(0.5, 0, 0.5)))

print(after::eatf(WWWusage, method = "EAT"))

print(after::eatf(WWWusage, method = "T"))


obj <- after::eatf(WWWusage, method = "EAT",
weights = c(0, 0.5, 0.5),
ci = "T")
plot(obj)


obj <- after::eatf(WWWusage, method = "EAT",
weights = c(0, 0.5, 0.5))
plot(obj)

}
