% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils_after.R
\name{embed_reg}
\alias{embed_reg}
\title{Embed time series}
\usage{
embed_reg(
  x,
  y,
  lags = 1,
  encoding = c("None", "model_matrix", "target_based"),
  testing = FALSE
)
}
\arguments{
\item{model_matrix}{}
}
\description{
Embed time series
}
\examples{

x <- 1:10
set.seed(2020)
df <- data.frame(x1 = x,
                 x2 = c("a", "b", "c", "b", "a",
                        "a", "a", "c", "b", "c"),
                 x3=sort(rnorm(10)))
df$x2 <- as.factor(df$x2)

lags <- 3L

cat("1 - data -----")
cat("\n")
print(df)
cat("\n")

cat("2 - training (model_matrix==FALSE) -----")
cat("\n")
cat("\n")
print(after::embed_reg(y=df$x1, x=df[, c("x2", "x3")],
                       lags = lags))
cat("\n")

cat("2 - training (model_matrix==TRUE) -----")
cat("\n")
cat("\n")
print(after::embed_reg(y=df$x1, x=df[, c("x2", "x3")],
                       lags = lags, encoding = "model_matrix"))
cat("\n")

cat("3 - training target-based -----")
cat("\n")
cat("\n")
print(after::embed_reg(y=df$x1, x=df[, c("x2", "x3")],
                       lags = lags, encoding = "target_based"))
cat("\n")

}
