% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_forecast_after_interface.R
\name{ensemblef}
\alias{ensemblef}
\title{Ensemble forecasting from fitted forecasting objects}
\usage{
ensemblef(
  fitted_objects,
  weights = rep(1/length(fitted_objects), length(fitted_objects)),
  ci = c("E", "A", "T", "garch", "gaussian")
)
}
\arguments{
\item{fitted_objects}{lorem ipsum}

\item{weights}{lorem ipsum}

\item{ci}{lorem ipsum}

\item{level}{lorem ipsum}
}
\description{
Ensemble forecasting from fitted forecasting objects
}
\examples{

x <- ts(rnorm(25))

obj1 <- forecast::thetaf(x)
obj2 <- forecast::forecast(forecast::ets(x))

list_obj <- list(obj1, obj2)

# par(mfrow=c(2, 2))
# plot(ensemblef(list_obj, ci="E"))
# plot(ensemblef(list_obj, ci="A"))
# plot(ensemblef(list_obj, ci="T"))
# plot(ensemblef(list_obj, ci="garch"))

}
