% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_forecast_after_interface.R
\name{garch11f}
\alias{garch11f}
\title{GARCH(1, 1) forecasting function}
\usage{
garch11f(x, h = 5, level = c(80, 95))
}
\arguments{
\item{x}{a univariate time series}

\item{h}{number of periods for forecasting}

\item{level}{confidence levels for prediction intervals}
}
\value{

}
\description{
GARCH(1, 1) forecasting function
}
\examples{

x <- ts(fGarch::garchSim()$garch)
# (obj <- garch11f(x))

}
