% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_correlated_var}
\alias{get_correlated_var}
\title{Simulate a correlated variable with a given correlation}
\usage{
get_correlated_var(x, rho, sim_fun = rnorm, seed = 123)
}
\arguments{
\item{x}{response variable}

\item{rho}{target correlation}

\item{sim_fun}{simulation function}
}
\value{

}
\description{
Simulate a correlated variable with a given correlation
}
\examples{

B <- 250
res <- rep(0, B)

for (i in 1:250)
{
  y <- rnorm(1000)
  res[i] <- get_correlated_var(y, rho = 0.1)$cors
}

hist(res)

}
