% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/88_utils.R
\name{get_uncertainty}
\alias{get_uncertainty}
\title{Title}
\usage{
get_uncertainty(
  mean_forecast,
  resids,
  h,
  level,
  x,
  ci = c("E", "A", "T", "garch")
)
}
\arguments{
\item{mean_forecast}{lorem ipsum}

\item{resids}{lorem ipsum}

\item{h}{lorem ipsum}

\item{level}{lorem ipsum}

\item{x}{lorem ipsum}

\item{ci}{lorem ipsum}
}
\description{
Title
}
\examples{

x <- AirPassengers
mean_forecast <- forecast::thetaf(x)$mean
resids <- forecast::thetaf(x)$resid
(res <- get_uncertainty(mean_forecast, resids, h=5, level=c(80, 95), x))
print(str(res))

}
