% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_forecast_mts_interface.R
\name{glmboostf}
\alias{glmboostf}
\title{Title}
\usage{
glmboostf(
  x,
  B = 10,
  eta = 0.1,
  lags = 1,
  nb_hidden = 1,
  nodes_sim = c("sobol", "halton", "unif"),
  activ = c("relu", "sigmoid", "tanh", "leakyrelu", "elu", "linear"),
  hidden_layer_bias = FALSE,
  a = 0.01,
  direct_link = FALSE,
  seed = 1,
  h = 5,
  type_ci = "none",
  type_forecast = c("recursive", "direct"),
  level = 95
)
}
\arguments{
\item{level}{}
}
\description{
Title
}
