% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_forecast_after_interface.R
\name{mmf}
\alias{mmf}
\title{Title}
\usage{
mmf(
  y,
  h = 5,
  level = c(80, 95),
  method = c("mean", "median"),
  fan = FALSE,
  lambda = NULL,
  biasadj = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{

par(mfrow=c(2, 2))
plot(mmf(Nile))
plot(mmf(Nile, method = "median"))
plot(mmf(AirPassengers, method = "median"))
plot(mmf(AirPassengers))

}
