% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_forecast_mts_interface.R
\name{pcrf}
\alias{pcrf}
\title{Title}
\usage{
pcrf(
  x,
  ncomp = ncol(x),
  lags = 1,
  h = 5,
  type_ci = "none",
  type_forecast = c("recursive", "direct"),
  level = 95
)
}
\arguments{
\item{level}{}
}
\description{
Title
}
