% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_forecast_mts_interface.R
\name{plsf}
\alias{plsf}
\title{Title}
\usage{
plsf(
  x,
  B = 5,
  lags = 1,
  nb_hidden = 5,
  nodes_sim = c("sobol", "halton", "unif"),
  activ = c("relu", "sigmoid", "tanh", "leakyrelu", "elu", "linear"),
  hidden_layer_bias = FALSE,
  direct_link = FALSE,
  a = 0.01,
  seed = 1,
  h = 5,
  type_ci = "none",
  type_forecast = c("recursive", "direct"),
  level = 95
)
}
\arguments{
\item{level}{}
}
\description{
Title
}
