% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_forecast_after_interface.R
\name{polythetaf}
\alias{polythetaf}
\title{Title}
\usage{
polythetaf(
  y,
  h = 5,
  lambda = 10^seq(from = -5, to = 4, length.out = 100),
  level = c(80, 95),
  fan = FALSE
)
}
\arguments{
\item{h}{lorem ipsum}

\item{lambda}{lorem ipsum}

\item{level}{lorem ipsum}

\item{fan}{lorem ipsum}

\item{x}{lorem ipsum}

\item{degree}{lorem ipsum}
}
\description{
Title
}
