% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_ridge.R
\name{predict_ridge}
\alias{predict_ridge}
\title{Title}
\usage{
predict_ridge(obj, newx, cv = TRUE)
}
\arguments{
\item{newx}{}
}
\value{

}
\description{
Title
}
\examples{


n <- 10000 ; p <- 100
X <- matrix(rnorm(n * p), n, p) # no intercept!
y <- rnorm(n)

fit_obj <- after::fit_ridge(X, y)

n_test <- 10

predict_ridge(fit_obj, newx=matrix(rnorm(n_test * p), n_test, p),
cv=TRUE)

predict_ridge(fit_obj, newx=matrix(rnorm(n_test * p), n_test, p),
cv=FALSE)


}
