% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_forecast_mts_interface.R
\name{ridge2f}
\alias{ridge2f}
\title{Title}
\usage{
ridge2f(
  y = x,
  lags = 1,
  nb_hidden = 5,
  nodes_sim = c("sobol", "halton", "unif"),
  activ = c("relu", "sigmoid", "tanh", "leakyrelu", "elu", "linear"),
  hidden_layer_bias = FALSE,
  col_sample = 1,
  row_sample = 1,
  a = 0.01,
  lambda_1 = 0.1,
  lambda_2 = 0.1,
  seed = 1,
  h = 5,
  type_ci = "none",
  type_forecast = c("recursive", "direct"),
  level = 95
)
}
\arguments{
\item{level}{}
}
\description{
Title
}
