% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_forecast_mts_interface.R
\name{ridgef}
\alias{ridgef}
\title{Title}
\usage{
ridgef(
  x,
  lags = 1,
  nb_hidden = 5,
  fit_method = c("ridge", "mgaussian"),
  nodes_sim = c("sobol", "halton", "unif"),
  activ = c("relu", "sigmoid", "tanh", "leakyrelu", "elu", "linear"),
  hidden_layer_bias = FALSE,
  a = 0.01,
  lambda = 0.1,
  alpha = 0.5,
  seed = 1,
  h = 5,
  type_ci = "none",
  type_forecast = c("recursive", "direct"),
  level = 95
)
}
\arguments{
\item{level}{}
}
\description{
Title
}
