% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale.ts}
\alias{scale.ts}
\title{Scale a univariate time series}
\usage{
\method{scale}{ts}(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{scale}{}
}
\value{

}
\description{
Scale a univariate time series
}
