% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils_after.R
\name{scale_after}
\alias{scale_after}
\title{Scale a univariate time series}
\usage{
scale_after(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{scale}{}
}
\description{
Scale a univariate time series
}
