% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_forecast_mts_interface.R
\name{scnf}
\alias{scnf}
\title{Title}
\usage{
scnf(
  x,
  B = 100,
  lags = 1,
  activ = c("tanh", "sigmoid"),
  hidden_layer_bias = FALSE,
  eta = 0.1,
  lam = 100,
  r = 0.3,
  tol = 1e-10,
  col_sample = 1,
  method = c("greedy", "direct"),
  type_optim = c("nlminb", "nmkb"),
  verbose = FALSE,
  h = 5,
  type_ci = "none",
  type_forecast = c("recursive", "direct"),
  level = 95
)
}
\arguments{
\item{level}{}
}
\description{
Title
}
