% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_forecast_after_interface.R
\name{smoothf}
\alias{smoothf}
\title{Title}
\usage{
smoothf(
  y,
  h = 5,
  level = c(80, 95),
  n_estimators = 5,
  lags = 1,
  method = c("mean", "median"),
  ci = c("E", "A", "T", "garch")
)
}
\arguments{
\item{n_estimators}{lorem ipsum}

\item{lags}{lorem ipsum}
}
\description{
Title
}
\examples{

par(mfrow=c(3, 2))

y <- Nile
(res <- smoothf(y, h=10))
plot(res)

y <- AirPassengers
(res <- smoothf(y, h = 10))
plot(res)

y <- USAccDeaths
(res <- smoothf(y, h = 10))
plot(res)

y <- WWWusage
(res <- smoothf(y, h = 10))
plot(res)

y <- WWWusage
(res <- smoothf(y, h = 10, lags = 2L))
plot(res)


y <- WWWusage
(res <- smoothf(y, h = 10, lags = 3L))
plot(res)


grid_params <- expand.grid(n_estimators=1:10, lags=c(1, 2, 3, 4))


}
