% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_forecast_mts_interface.R
\name{stack_data_ridge2f}
\alias{stack_data_ridge2f}
\title{Title}
\usage{
stack_data_ridge2f(
  x,
  B = 10,
  lags = 1,
  nb_hidden = 5,
  nodes_sim = c("sobol", "halton", "unif"),
  activ = c("relu", "sigmoid", "tanh", "leakyrelu", "elu", "linear"),
  hidden_layer_bias = FALSE,
  col_sample = 0.9,
  row_sample = 0.8,
  a = 0.01,
  lambda_1 = 0.1,
  lambda_2 = 0.1
)
}
\arguments{
\item{lambda_2}{}
}
\description{
Title
}
