% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{target_based_encoder}
\alias{target_based_encoder}
\title{Target-based encoder}
\usage{
target_based_encoder(x, y, rho = 0, seed = 123)
}
\arguments{
\item{x}{matrix; explanatory variables}

\item{y}{vector; response}

\item{rho}{float; desired correlation}
}
\value{
a list, encoded regressors and codes
}
\description{
Target-based encoder
}
\examples{

n <- 100
X <- cbind.data.frame(as.factor(sample(x = c(1, 2, 3),
size = n, replace = TRUE)), as.factor(sample(x = c(0, 1),
size = n, replace = TRUE)))
X$X3 <- rnorm(n)
colnames(X) <- c("X1", "X2", "X3")

y <- rt(nrow(X), df=2)

z <- after::target_based_encoder(X, y)$newx

head(X)
head(z)
tail(X)
tail(z)

cor(y, z[,2])
cor(y, z[,1])

sd(z[,2])
sd(z[,1])

}
