% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils_after.R
\name{treat_outliers}
\alias{treat_outliers}
\title{Outliers treatment}
\usage{
treat_outliers(x, plot = FALSE, replace = TRUE)
}
\arguments{
\item{x}{a time series}

\item{plot}{a boolean; plot time series or not}

\item{replace}{a boolean; replace outliers or not? (if TRUE, creates a new time series)}
}
\value{

}
\description{
Outliers treatment
}
\examples{

x <- rnorm(100)
x[5] <- 10
x[10] <- 10
x[20] <- 10
x[70] <- 10
after::treat_outliers(x, plot = TRUE)

}
