% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_forecast_mts_interface.R
\name{varf}
\alias{varf}
\title{Title}
\usage{
varf(
  x,
  lags = 1,
  penalization = c("none", "l1"),
  lambda = 0.1,
  type_VAR = c("const", "trend", "both", "none"),
  type_forecast = c("recursive", "direct"),
  h = 5,
  type_ci = "none",
  level = 95
)
}
\arguments{
\item{level}{}
}
\description{
Title
}
