% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{arimaf}
\alias{arimaf}
\title{Title}
\usage{
arimaf(
  x,
  h = 5,
  level = c(80, 95),
  ci = c("gaussian", "E", "A", "T", "garch"),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{

after::arimaf(AirPassengers, h=7)

after::arimaf(Nile, h=6)

after::arimaf(WWWusage, h=8)

after::arimaf(USAccDeaths, h=10)

}
