% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_dynrm.R
\name{dynrm_fit}
\alias{dynrm_fit}
\title{Fit a dynrm}
\usage{
dynrm_fit(
  y,
  p,
  P = 1,
  xreg = NULL,
  fit_func = stats::lm.fit,
  predict_func = predict.lm,
  fit_params = NULL,
  lambda = NULL,
  alpha = NULL,
  scale_inputs = TRUE,
  x = y,
  centers = NULL,
  seed = 123,
  ...
)
}
\description{
Fit a dynrm
}
