% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynrm.R
\name{dynrm_predict}
\alias{dynrm_predict}
\title{Predict from dynrm}
\usage{
dynrm_predict(
  out,
  h = ifelse(out$m > 1, 2 * out$m, 10),
  level = c(80, 95),
  fan = FALSE,
  xreg = NULL,
  lambda = out$lambda,
  ci = c("E", "A", "T", "garch", "gaussian"),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Predict from dynrm
}
\examples{

dat <- list(lynx, USAccDeaths, Nile, WWWusage, fdeaths, AirPassengers)

h <- 10

# Example 2: ridge

par(mfrow=c(3, 2))

for (i in 1:length(dat))
{
 x <- dat[[i]]
 cat("Dataset #", i, ":", length(x), "points \n")

 obj <- after::dynrm_fit(x, fit_func = after::fit_ridge,
 predict_func = after::predict_ridge, xreg = as.matrix(1:length(x)))

  plot(after::dynrm_predict(obj, xreg = as.matrix((length(x)+1):(length(x)+h-1)),
  ci="A", h=h))
}

}
