% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_forecast_after_interface.R
\name{dynrmf}
\alias{dynrmf}
\title{Title}
\usage{
dynrmf(
  y,
  h = 10,
  level = c(80, 95),
  fit_func = fit_ridge,
  predict_func = predict_ridge,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{

par(mfrow=c(3, 2))
plot(dynrmf(USAccDeaths, level=c(80, 90, 95, 99)))
plot(dynrmf(AirPassengers, level=c(80, 90, 95, 99)))
plot(dynrmf(lynx, level=c(80, 90, 95, 99)))
plot(dynrmf(WWWusage, level=c(80, 90, 95, 99)))
plot(dynrmf(Nile, level=c(80, 90, 95, 99)))
plot(dynrmf(fdeaths, level=c(80, 90, 95, 99)))


}
