% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polythetaf.R
\name{polythetaf}
\alias{polythetaf}
\title{Title}
\usage{
polythetaf(
  y,
  h = ifelse(frequency(x) > 1, 2 * frequency(x), 10),
  lambda = 10^seq(from = -5, to = 4, length.out = 100),
  level = c(80, 95),
  fan = FALSE
)
}
\arguments{
\item{fan}{}
}
\value{

}
\description{
Title
}
